Extracted gated blocks parsing and manipulation to content-gating service
no issue

- logic for handling of gated blocks lived in the post output serializer's "post gating" module which meant it was disjointed from other content gating related code and was difficult to re-use in other areas
- moved the logic to the new `content-gating` service to keep related functionality co-located and reusable
