+++ ghost/core/core/server/api/endpoints/utils/serializers/output/utils/post-gating.js
@@ -2,74 +2,6 @@ const contentGatingService = require('../../../../../../services/content-gating'
 const labs = require('../../../../../../../shared/labs');
 const htmlToPlaintext = require('@tryghost/html-to-plaintext');
 
-const {PERMIT_ACCESS} = contentGatingService;
-
-// Match the start of a gated block - fast regex as a pre-check before doing full regex+loop
-const HAS_GATED_BLOCKS_REGEX = /<!--\s*kg-gated-block:begin/;
-// Match gated block comments
-// e.g. <!--kg-gated-block:begin nonMember:true memberSegment:"status:free"-->...gated content<!--kg-gated-block:end-->
-const GATED_BLOCK_REGEX = /<!--\s*kg-gated-block:begin\s+([^\n]+?)\s*-->\s*([\s\S]*?)\s*<!--\s*kg-gated-block:end\s*-->/g;
-// Match the key-value pairs (with optional quotes around the value) in the gated-block:begin comment
-const GATED_BLOCK_PARAM_REGEX = /\b(?<key>\w+):["']?(?<value>[^"'\s]+)["']?/g;
-
-const ALLOWED_GATED_BLOCK_PARAMS = {
-    nonMember: {type: 'boolean'},
-    memberSegment: {type: 'string', allowedValues: ['', 'status:free,status:-free', 'status:free', 'status:-free']}
-};
-const ALLOWED_GATED_BLOCK_KEYS = Object.keys(ALLOWED_GATED_BLOCK_PARAMS);
-
-const parseGatedBlockParams = function (paramsString) {
-    const params = {};
-
-    const matches = paramsString.matchAll(GATED_BLOCK_PARAM_REGEX);
-    for (const match of matches) {
-        const key = match.groups.key;
-        let value = match.groups.value;
-
-        if (!ALLOWED_GATED_BLOCK_KEYS.includes(key)) {
-            continue;
-        }
-
-        // Convert "true"/"false" strings to booleans, otherwise keep as string
-        if (value === 'true') {
-            value = true;
-        } else if (value === 'false') {
-            value = false;
-        }
-
-        if (typeof value !== ALLOWED_GATED_BLOCK_PARAMS[key].type) {
-            continue;
-        }
-
-        if (ALLOWED_GATED_BLOCK_PARAMS[key].allowedValues && !ALLOWED_GATED_BLOCK_PARAMS[key].allowedValues.includes(value)) {
-            continue;
-        }
-
-        params[key] = value;
-    }
-
-    return params;
-};
-
-/**
- * @param {string} html - The HTML to strip gated blocks from
- * @param {object} member - The member who's access should be checked
- * @returns {string} HTML with gated blocks stripped
- */
-const stripGatedBlocks = function (html, member) {
-    return html.replace(GATED_BLOCK_REGEX, (match, params, content) => {
-        const gatedBlockParams = module.exports.parseGatedBlockParams(params);
-        const checkResult = contentGatingService.checkGatedBlockAccess(gatedBlockParams, member);
-
-        if (checkResult === PERMIT_ACCESS) {
-            // return content rather than match to avoid rendering gated block wrapping comments
-            return content;
-        } else {
-            return '';
-        }
-    });
-};
-
 function _updatePlaintext(attrs) {
     if (attrs.html) {
         attrs.plaintext = htmlToPlaintext.excerpt(attrs.html);
@@ -107,13 +39,10 @@ const forPost = (attrs, frame) => {
         }
     }
 
-    if (labs.isSet('contentVisibility')) {
-        const hasGatedBlocks = HAS_GATED_BLOCKS_REGEX.test(attrs.html);
-        if (hasGatedBlocks) {
-            attrs.html = module.exports.stripGatedBlocks(attrs.html, frame.original.context.member);
-            _updatePlaintext(attrs);
-            _updateExcerpt(attrs);
-        }
+    if (labs.isSet('contentVisibility') && contentGatingService.htmlHasGatedBlocks(attrs.html)) {
+        attrs.html = contentGatingService.removeGatedBlocksFromHtml(attrs.html, frame.original.context.member);
+        _updatePlaintext(attrs);
+        _updateExcerpt(attrs);
     }
 
     if (!Object.prototype.hasOwnProperty.call(frame.options, 'columns') || (frame.options.columns.includes('access'))) {
@@ -124,7 +53,5 @@ const forPost = (attrs, frame) => {
 };
 
 module.exports = {
-    parseGatedBlockParams,
-    stripGatedBlocks,
     forPost
 };
