+++ ghost/core/core/server/services/content-gating/index.js
@@ -1,5 +1,6 @@
 const {PERMIT_ACCESS, BLOCK_ACCESS} = require('./constants');
-const {checkPostAccess, checkGatedBlockAccess} = require('./access-checks');
+const accessChecks = require('./access-checks');
+const gatedBlocks = require('./gated-blocks');
 
 /**
  * @typedef {import ('./typings').GatedPost} GatedPost
@@ -9,6 +10,11 @@ const {checkPostAccess, checkGatedBlockAccess} = require('./access-checks');
 module.exports = {
     PERMIT_ACCESS,
     BLOCK_ACCESS,
-    checkPostAccess,
-    checkGatedBlockAccess
+    checkPostAccess: accessChecks.checkPostAccess,
+    checkGatedBlockAccess: accessChecks.checkGatedBlockAccess,
+    htmlHasGatedBlocks: gatedBlocks.htmlHasGatedBlocks,
+    // avoid circular require dependency by using "DI" for checkGatedBlockAccess
+    removeGatedBlocksFromHtml: (html, member) => {
+        return gatedBlocks.removeGatedBlocksFromHtml(html, member, accessChecks.checkGatedBlockAccess);
+    }
 };
